Thanks lord! Requirements working with default values!

+++ src/main/java/org/fogbowcloud/manager/core/plugins/compute/opennebula/OpenNebulaComputePlugin.java
@@ -707,7 +707,7 @@ public class OpenNebulaComputePlugin implements ComputePlugin {
 			String term, int minimumValue) {
 		String valueStr = RequirementsHelper
 				.getSmallestValueForAttribute(requirements, term);
-		return String.valueOf(Math.max(Double.valueOf(valueStr), minimumValue));
+		return String.valueOf(Math.max(Integer.valueOf(valueStr), minimumValue));
 	}
 
 	protected List<String> getTemplatesInProperties(Properties properties) {
